feat: Перевел проект на FSD
