+++ package.json
@@ -15,6 +15,7 @@
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
+    "@a1rth/css-normalize": "^1.0.9",
     "@million/lint": "latest",
     "@reduxjs/toolkit": "^2.2.3",
     "@types/bun": "latest",
