+++ project_structure.txt
@@ -1,0 +1,103 @@
+.
+├── README.md
+├── bun.lockb
+├── index.html
+├── package.json
+├── project_structure.txt
+├── public
+│   └── vite.svg
+├── src
+│   ├── app
+│   │   ├── App.module.css
+│   │   ├── App.tsx
+│   │   ├── index.css
+│   │   ├── index.tsx
+│   │   ├── layout
+│   │   │   └── Root
+│   │   │       └── Root.tsx
+│   │   └── providers
+│   │       ├── router
+│   │       │   └── routes.tsx
+│   │       └── store
+│   │           ├── localStorage.ts
+│   │           ├── movies.slice.ts
+│   │           ├── store.ts
+│   │           └── users.slice.ts
+│   ├── assets
+│   │   ├── icons
+│   │   │   ├── like.svg
+│   │   │   ├── logo.svg
+│   │   │   ├── search.svg
+│   │   │   ├── star.svg
+│   │   │   └── trash.svg
+│   │   └── no-image-placeholder.png
+│   ├── pages
+│   │   ├── Error
+│   │   │   ├── Error.module.css
+│   │   │   ├── Error.props.ts
+│   │   │   └── Error.tsx
+│   │   ├── Favorites
+│   │   │   └── Favorites.tsx
+│   │   ├── Feed
+│   │   │   ├── Feed.module.css
+│   │   │   └── Feed.tsx
+│   │   ├── Login
+│   │   │   ├── Login.module.css
+│   │   │   └── Login.tsx
+│   │   └── Movie
+│   │       ├── Movie.module.css
+│   │       └── Movie.tsx
+│   ├── shared
+│   │   ├── components
+│   │   │   ├── Button
+│   │   │   │   ├── Button.module.css
+│   │   │   │   ├── Button.props.ts
+│   │   │   │   └── Button.tsx
+│   │   │   ├── Heading
+│   │   │   │   ├── Heading.module.css
+│   │   │   │   ├── Heading.props.ts
+│   │   │   │   └── Heading.tsx
+│   │   │   ├── Input
+│   │   │   │   ├── Input.module.css
+│   │   │   │   ├── Input.props.ts
+│   │   │   │   └── Input.tsx
+│   │   │   ├── ItemList
+│   │   │   │   ├── ItemList.module.css
+│   │   │   │   ├── ItemList.props.ts
+│   │   │   │   └── ItemList.tsx
+│   │   │   ├── Navbar
+│   │   │   ├── Paragraph
+│   │   │   │   ├── Paragraph.module.css
+│   │   │   │   ├── Paragraph.props.ts
+│   │   │   │   └── Paragraph.tsx
+│   │   │   └── Rating
+│   │   │       ├── Rating.module.css
+│   │   │       └── Rating.tsx
+│   │   ├── helpers
+│   │   │   └── RequireAuth.tsx
+│   │   ├── hooks
+│   │   │   ├── useApi.ts
+│   │   │   ├── useAppNavigation.ts
+│   │   │   └── useAuth.ts
+│   │   └── interfaces
+│   │       ├── movie.interface.ts
+│   │       └── user.interface.ts
+│   ├── types
+│   │   └── index.d.ts
+│   └── widgets
+│       ├── Favorite
+│       │   ├── Favorite.module.css
+│       │   ├── Favorite.props.ts
+│       │   └── Favorite.tsx
+│       ├── Item
+│       │   ├── Item.module.css
+│       │   ├── Item.props.ts
+│       │   └── Item.tsx
+│       └── Navbar
+│           ├── Navbar.module.css
+│           └── Navbar.tsx
+├── tsconfig.json
+├── tsconfig.node.json
+└── vite.config.ts
+
+34 directories, 67 files
