+++ src/app/index.tsx
@@ -1,4 +1,6 @@
 import ReactDOM from 'react-dom/client';
 import App from './App';
+import './index.css';
+import '@a1rth/css-normalize';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(<App />);
