+++ src/app/layout/Root/Root.tsx
@@ -1,4 +1,4 @@
-import Navbar from '@/components/Navbar/Navbar';
+import Navbar from '@/widgets/Navbar/Navbar';
 import { Outlet } from 'react-router-dom';
 const Root = () => {
   return (
