--- src/components/Feedback/Feedback.tsx
@@ -1,7 +1,0 @@
-import styles from './Feedback.module.css';
-
-const Feedback = ({}) => {
-  return <div>Feedback</div>;
-};
-
-export default Feedback;
