--- src/helpers/data.ts
@@ -1,57 +1,0 @@
-export const movies = [
-  {
-    id: 1,
-    name: "Black Window",
-    rating: 324,
-    image: "https://i.imgur.com/dbzIjPO.png",
-  },
-
-  {
-    id: 2,
-    name: "Shang Chi",
-    rating: 124,
-    image: "https://i.imgur.com/dbzIjPO.png",
-  },
-
-  {
-    id: 3,
-    name: "Loki",
-    rating: 235,
-    image: "https://i.imgur.com/dbzIjPO.png",
-  },
-
-  {
-    id: 4,
-    name: "How I Met Your Mother",
-    rating: 123,
-    image: "https://i.imgur.com/dbzIjPO.png",
-  },
-
-  {
-    id: 5,
-    name: "Money Heist",
-    rating: 8125,
-    image: "https://i.imgur.com/dbzIjPO.png",
-  },
-
-  {
-    id: 6,
-    name: "Friends",
-    rating: 123,
-    image: "https://i.imgur.com/dbzIjPO.png",
-  },
-
-  {
-    id: 7,
-    name: "The Big Bang Theory",
-    rating: 12,
-    image: "https://i.imgur.com/dbzIjPO.png",
-  },
-
-  {
-    id: 8,
-    name: "Two And a Half Men",
-    rating: 456,
-    image: "https://i.imgur.com/dbzIjPO.png",
-  },
-];
