+++ src/pages/Login/Login.tsx
@@ -1,14 +1,14 @@
-import Input from '@/components/Input/Input';
+import Input from '@/shared/components/Input/Input';
 import styles from './Login.module.css';
-import Button from '@/components/Button/Button';
-import Heading from '@/components/Heading/Heading';
+import Button from '@/shared/components/Button/Button';
+import Heading from '@/shared/components/Heading/Heading';
 import { useContext, useRef } from 'react';
 import { UserContext } from '@/context/user.context';
-import { useAppNavigation } from '@/hooks/useAppNavigation';
+import { useAppNavigation } from '@/shared/hooks/useAppNavigation';
 import { useDispatch } from 'react-redux';
-import type { AppDispatch } from '@/store/store';
+import type { AppDispatch } from '@/app/providers/store/store';
 
-import { login } from '@/store/users.slice';
+import { login } from '@/app/providers/store/users.slice';
 
 const Login = () => {
   const inputRef = useRef<HTMLInputElement>(null);
