+++ src/shared/components/Heading/Heading.props.ts
