+++ src/shared/components/Paragraph/Paragraph.props.ts
