+++ src/shared/components/Rating/Rating.module.css
