+++ src/shared/components/Rating/Rating.tsx
