+++ src/shared/interfaces/movie.interface.ts
