+++ src/widgets/Favorite/Favorite.module.css
