+++ src/widgets/Favorite/Favorite.props.ts
