+++ src/widgets/Item/Item.tsx
@@ -3,7 +3,7 @@ import styles from './Item.module.css';
 import type { ItemProps } from './Item.props';
 import { Link } from 'react-router-dom';
 
-import Rating from '../Rating/Rating';
+import Rating from '../../shared/components/Rating/Rating';
 import Favorite from '../Favorite/Favorite';
 
 const Item = ({ id, name, image, rating }: ItemProps) => {
